requirements.txt
asgiref==3.6.0
axios==0.4.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Django==3.2
django-ckeditor==6.5.1
django-cors-headers==4.2.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.0.0
idna==3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
Pillow==9.5.0
Pygments==2.15.1
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
rich==13.4.2
sqlparse==0.4.4
urllib3==2.0.4
whitenoise==6.5.0
